refactor: Change class name - MailService -> MailComponent - 임시 비밀번호 전송 폼 작성
