+++ src/main/java/com/eunsun/travel_mate/component/MailComponent.java
@@ -1,4 +1,4 @@
-package com.eunsun.travel_mate.service;
+package com.eunsun.travel_mate.component;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.mail.SimpleMailMessage;
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class MailService {
+public class MailComponent {
 
   private final JavaMailSender mailSender;
 
@@ -22,6 +22,12 @@ public class MailService {
   public void sendVerificationEmail(String to, String verificationCode) {
     String subject = "Travel Mate 회원가입 인증 코드";
     String text = "인증 코드: " + verificationCode;
+    sendEmail(to, subject, text);
+  }
+
+  public void sendTemporaryPasswordEmail(String to, String temporaryPassword) {
+    String subject = "Travel Mate 임시 비밀번호";
+    String text = "임시 비밀번호: " + temporaryPassword;
     sendEmail(to, subject, text);
   }
 }
